Misc stuff

+++ src/minirds.c
@@ -40,7 +40,7 @@ static inline void float2char2channel(
 	int8_t lower, upper;
 
 	for (uint16_t i = 0; i < frames; i++) {
-		sample = lroundf((float)(inbuf[j] + inbuf[j+1]) * 16383.5f);
+		sample = lroundf((inbuf[j] + inbuf[j+1]) * 16383.5f);
 
 		// convert from short to char
 		lower = sample & 255;
